:recycle: moved Color class
