+++ pincer/utils/__init__.py
@@ -2,6 +2,7 @@
 # Full MIT License can be found in `LICENSE` at the project root.
 
 from .api_object import APIObject, HTTPMeta
+from .color import Color
 from .conversion import convert
 from .event_mgr import EventMgr
 from .directory import chdir
@@ -19,7 +20,7 @@ from .types import (
 
 
 __all__ = (
-    "APINullable", "APIObject", "CheckFunction", "Choices", "Coro",
+    "APINullable", "APIObject", "CheckFunction", "Choices", "Coro", "Color",
     "Descripted", "EventMgr", "HTTPMeta", "MISSING", "MissingType",
     "Snowflake", "Task", "TaskScheduler", "Timestamp", "chdir",
     "choice_value_types", "convert", "get_index", "get_params",
